Fixes for a couple of minor compiler warnings

+++ modules/juce_osc/osc/juce_OSCReceiver.cpp
@@ -234,7 +234,7 @@ namespace
                 bundle.addElement (readElement());
 
                 auto newPos = getPosition();
-                bytesRead += newPos - pos;
+                bytesRead += (size_t) (newPos - pos);
                 pos = newPos;
             }
